+++ src/Protocol/CallService.php
@@ -77,10 +77,7 @@ class CallService
                 $inputArgumentResults[] = $decoder->readUInt32();
             }
 
-            $diagCount = $decoder->readInt32();
-            for ($j = 0; $j < $diagCount; $j++) {
-                $this->skipDiagnosticInfo($decoder);
-            }
+            $decoder->skipDiagnosticInfoArray();
 
             $outputArgCount = $decoder->readInt32();
             for ($j = 0; $j < $outputArgCount; $j++) {
@@ -88,10 +85,7 @@ class CallService
             }
         }
 
-        $diagCount = $decoder->readInt32();
-        for ($i = 0; $i < $diagCount; $i++) {
-            $this->skipDiagnosticInfo($decoder);
-        }
+        $decoder->skipDiagnosticInfoArray();
 
         return new CallResult($statusCode, $inputArgumentResults, $outputArguments);
     }
@@ -151,32 +145,6 @@ class CallService
         $body->writeInt32(count($inputArguments));
         foreach ($inputArguments as $arg) {
             $body->writeVariant($arg);
-        }
-    }
-
-    /**
-     * @param BinaryDecoder $decoder
-     */
-    private function skipDiagnosticInfo(BinaryDecoder $decoder): void
-    {
-        $mask = $decoder->readByte();
-        if ($mask & 0x01) {
-            $decoder->readInt32();
-        }
-        if ($mask & 0x02) {
-            $decoder->readInt32();
-        }
-        if ($mask & 0x04) {
-            $decoder->readInt32();
-        }
-        if ($mask & 0x08) {
-            $decoder->readString();
-        }
-        if ($mask & 0x10) {
-            $decoder->readUInt32();
-        }
-        if ($mask & 0x20) {
-            $this->skipDiagnosticInfo($decoder);
         }
     }