+++ src/Protocol/HistoryReadService.php
@@ -162,10 +162,7 @@ class HistoryReadService
             }
         }
 
-        $diagCount = $decoder->readInt32();
-        for ($i = 0; $i < $diagCount; $i++) {
-            $this->skipDiagnosticInfo($decoder);
-        }
+        $decoder->skipDiagnosticInfoArray();
 
         return $allValues;
     }
@@ -320,32 +317,6 @@ class HistoryReadService
         $body->writeUInt32(10000);
         $body->writeNodeId(NodeId::numeric(0, 0));
         $body->writeByte(0);
-    }
-
-    /**
-     * @param BinaryDecoder $decoder
-     */
-    private function skipDiagnosticInfo(BinaryDecoder $decoder): void
-    {
-        $mask = $decoder->readByte();
-        if ($mask & 0x01) {
-            $decoder->readInt32();
-        }
-        if ($mask & 0x02) {
-            $decoder->readInt32();
-        }
-        if ($mask & 0x04) {
-            $decoder->readInt32();
-        }
-        if ($mask & 0x08) {
-            $decoder->readString();
-        }
-        if ($mask & 0x10) {
-            $decoder->readUInt32();
-        }
-        if ($mask & 0x20) {
-            $this->skipDiagnosticInfo($decoder);
-        }
     }
 
     /**