+++ src/Protocol/MonitoredItemService.php
@@ -81,10 +81,7 @@ class MonitoredItemService
             $results[] = new MonitoredItemResult($statusCode, $monitoredItemId, $revisedSamplingInterval, $revisedQueueSize);
         }
 
-        $diagCount = $decoder->readInt32();
-        for ($i = 0; $i < $diagCount; $i++) {
-            $this->skipDiagnosticInfo($decoder);
-        }
+        $decoder->skipDiagnosticInfoArray();
 
         return $results;
     }
@@ -337,10 +334,7 @@ class MonitoredItemService
             $results[] = $decoder->readUInt32();
         }
 
-        $diagCount = $decoder->readInt32();
-        for ($i = 0; $i < $diagCount; $i++) {
-            $this->skipDiagnosticInfo($decoder);
-        }
+        $decoder->skipDiagnosticInfoArray();
 
         return $results;
     }
@@ -404,32 +398,6 @@ class MonitoredItemService
         $body->writeUInt32(10000);
         $body->writeNodeId(NodeId::numeric(0, 0));
         $body->writeByte(0);
-    }
-
-    /**
-     * @param BinaryDecoder $decoder
-     */
-    private function skipDiagnosticInfo(BinaryDecoder $decoder): void
-    {
-        $mask = $decoder->readByte();
-        if ($mask & 0x01) {
-            $decoder->readInt32();
-        }
-        if ($mask & 0x02) {
-            $decoder->readInt32();
-        }
-        if ($mask & 0x04) {
-            $decoder->readInt32();
-        }
-        if ($mask & 0x08) {
-            $decoder->readString();
-        }
-        if ($mask & 0x10) {
-            $decoder->readUInt32();
-        }
-        if ($mask & 0x20) {
-            $this->skipDiagnosticInfo($decoder);
-        }
     }
 
     /**